Changed values to float and refactored translate command

+++ mobile/app/src/main/java/com/example/colorimagemobile/services/drawing/TransformationManager.kt
@@ -7,7 +7,7 @@ import com.example.colorimagemobile.classes.toolsCommand.RectangleCommand
 object TransformationManager {
     var previousTransformation: HashMap<String, String> = HashMap()
 
-    fun saveTranslateTransformation(deltaX: Int, deltaY: Int, id: String, transformationLog: String, shapeLabel: ShapeLabel){
+    fun saveTranslateTransformation(deltaX: Float, deltaY: Float, id: String, transformationLog: String, shapeLabel: ShapeLabel){
         val currentTransformation = " translate($deltaX $deltaY)"
         previousTransformation[id] = currentTransformation + transformationLog
